Fix category multiselect in search results nav
The search results page was loading a group of `virtualTypes` which were
not respecting the underlying class preferences or plugins defined in
the layered navigation module. This commit resolves this issue.

This commit also sets the root category of the current store as a filter
when building the general and partial results buckets. This is done to
correct the filter item counts returned when multiple category selection
are active. The `current_category` registry value is used for
determining if the request is site search or a category view. It is
assumed to be site search if the `current_category` is not set.